feat(kernel-env): extract environment management into shared crate
Create a new `kernel-env` crate that encapsulates all Python environment
management (UV + Conda) with a progress reporting trait. This enables
rich progress feedback on the daemon path and makes the environment
logic reusable across notebook, runtimed, and tauri-jupyter.

Key changes:
- New `kernel-env` crate with `ProgressHandler` trait, `EnvProgressPhase`
  enum, and `RattlerReporter` (rattler's `Reporter` impl)
- UV and Conda environment creation, caching, prewarming all moved to
  kernel-env with progress callbacks throughout
- `runtimed/inline_env.rs` now delegates to kernel-env with a
  `BroadcastProgressHandler` that forwards progress via broadcast channel
- `notebook/conda_env.rs` and `notebook/uv_env.rs` become thin wrappers
  keeping only notebook-specific metadata operations (nbformat)
- Added `EnvProgress` variant to `NotebookBroadcast` protocol for
  daemon-to-frontend progress events
- All environment functions accept `Arc<dyn ProgressHandler>` for
  flexible progress routing (Tauri events, broadcast, logs)
