+++ crates/kernel-env/Cargo.toml
@@ -1,0 +1,37 @@
+[package]
+name = "kernel-env"
+version = "0.1.0"
+edition = "2021"
+description = "Python environment management (UV + Conda) with progress reporting"
+repository = "https://github.com/runtimed/runt"
+license = "BSD-3-Clause"
+
+[lints]
+workspace = true
+
+[dependencies]
+anyhow = { workspace = true }
+tokio = { workspace = true }
+serde = { workspace = true }
+serde_json = { workspace = true }
+uuid = { workspace = true }
+log = "0.4"
+dirs = "5"
+sha2 = "0.10"
+hex = "0.4"
+
+# Tool bootstrapping (get_uv_path)
+kernel-launch = { path = "../kernel-launch" }
+
+# Conda environment support via rattler
+rattler = "0.39"
+rattler_cache = "0.6"
+rattler_conda_types = "0.43"
+rattler_repodata_gateway = { version = "0.25", features = ["gateway"] }
+rattler_solve = { version = "4.2", features = ["resolvo"] }
+rattler_virtual_packages = "2.3"
+reqwest = { version = "0.12", default-features = false, features = ["rustls-tls"] }
+reqwest-middleware = "0.4"
+
+[dev-dependencies]
+tempfile = "3"
