+++ crates/notebook/Cargo.toml
@@ -35,6 +35,7 @@ tauri-jupyter = { path = "../tauri-jupyter" }
 runtimed = { path = "../runtimed" }
 runt-trust = { path = "../runt-trust" }
 kernel-launch = { path = "../kernel-launch" }
+kernel-env = { path = "../kernel-env" }
 nbformat = "1.2.0"
 petname = "2"
 log = "0.4"
